Se ha terminado de arreglar el juego de Loteria
También se han arreglado varios errores. Ver especialmente la clase WindowLoteria.java

+++ LaMiaFortuna/src/main/java/clases/Ficha.java
@@ -13,7 +13,7 @@ public class Ficha extends CosaComprable {
 
     private String color;
 
-    public Ficha(String color, int precio) {
+    public Ficha(String color, double precio) {
         super(precio);
         this.color = color;
     }
